fix: run FEDformer FFT in float32 under AMP

+++ baselines/FEDformer/layers/AutoCorrelation.py
@@ -124,8 +124,14 @@ class AutoCorrelation(nn.Module):
             keys = keys[:, :L, :, :]
 
         # period-based dependencies
-        q_fft = torch.fft.rfft(queries.permute(0, 2, 3, 1).contiguous(), dim=-1)
-        k_fft = torch.fft.rfft(keys.permute(0, 2, 3, 1).contiguous(), dim=-1)
+        q_fft = torch.fft.rfft(
+            queries.permute(0, 2, 3, 1).contiguous().float(),
+            dim=-1,
+        )
+        k_fft = torch.fft.rfft(
+            keys.permute(0, 2, 3, 1).contiguous().float(),
+            dim=-1,
+        )
         res = q_fft * torch.conj(k_fft)
         corr = torch.fft.irfft(res, dim=-1)
